Extract common test stuff.
